% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\docType{data}
\name{geno}
\alias{geno}
\title{The genotype data}
\format{A matrix with 100 row and 3 variables
\describe{
\item{AA}{the number of the individuals with genotype AA}
\item{Aa}{the number of the individuals with genotype Aa}
\item{Aa}{the number of the individuals with genotype aa}
}}
\usage{
geno
}
\description{
The dataset of the samples from multinomial distribution.
}
\keyword{datasets}
