% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{maf_calculation}
\alias{maf_calculation}
\title{Calculation of minor allele frequency}
\usage{
maf_calculation(geno, comment = TRUE)
}
\arguments{
\item{geno}{vector, the number of three genotypes AA, Aa, aa}

\item{comment}{a logical value indicating whether the comment "Hellow!" is shown. default is TRUE.}
}
\value{
Minor allele frequency of the locus
}
\description{
This Function is to calculate minor allele frequency.library
}
\examples{
library(mysample)
geno <- c(22,56,22)
maf_calculation(N)

}
\keyword{NA}
